+++ packages/common/src/types/context.ts
@@ -116,6 +116,7 @@ export interface Group {
   color?: string;
   finished?: boolean;
   interruptId?: string;
+  ephemeral?: boolean;
 }
 
 export interface UsageReportData {